+++ CHANGELOG.md
@@ -5,11 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+Added / Changed / Deprecated / Removed / Fixed / Security
+
 ## [Unreleased]
 
 > Corresponds to changes in the `develop` branch since the last release
 
-Added / Changed / Deprecated / Removed / Fixed / Security
+### Added
+
+#### org.ojalgo.function
+
+- New special functions: beta (complete, incomplete and regularized), gamma (logarithmic), Hypergeometric and Pochhammer symbol. Inluding complex valued variants where applicable.
+
+#### org.ojalgo.random
+
+- The `getDistribution()` method in the TDistribution is now implemented for the general case. Previously it was only implemented for a few distinct degrees of freedom.
 
 ## [47.3.1] – 2019-09-29